add publication date
also consolidates utils
