+++ src/layouts/Layout.astro
@@ -1,6 +1,7 @@
 ---
 import "@fontsource/source-serif-pro";
-import { basePath, siteName, descriptionPrefix } from "../utils";
+import { basePath } from "../utils";
+import { siteName, descriptionPrefix } from "../utils/meta";
 import { ViewTransitions } from "astro:transitions";
 interface Props {
   isIndexPage?: boolean;
