+++ src/pages/dg/[gd_rule].astro
@@ -4,7 +4,7 @@ import H1 from "../../components/H1.astro";
 import H2 from "../../components/H2.astro";
 import Link from "../../components/Link.astro";
 import RuleBodyContent from "../../components/RuleBodyContent.astro";
-import getAllBookRuleData from "../../utils/getAllBookRuleData";
+import { getAllBookRuleData } from "../../utils";
 import { getSchoolPathFilePath } from "../../utils";
 
 import { schoolBookGroups } from "../../data/content";
