+++ src/utils/meta.ts
@@ -1,0 +1,4 @@
+export const siteName = "Translations of the Chinese Bhikkhunī Vinayas";
+export const descriptionPrefix =
+  "Vimalañāṇī Bhikkhunī’s English translation of ";
+export const dateFirstPublished = "2024-06-05";
